tests/wpt/web-platform-tests/infrastructure/expected-fail/uncaught-exception-following-subtest.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Uncaught exception following subtest</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {});
throw new Error("error outside any setup or test");
</script>
