LayoutTests/imported/w3c/web-platform-tests/infrastructure/expected-fail/uncaught-exception-following-subtest-expected.txt

CONSOLE MESSAGE: Error: error outside any setup or test

Harness Error (FAIL), message = Error: error outside any setup or test

PASS Uncaught exception following subtest
